Raise upper bound on cryptography dependency
Currently python-cryptography in Rawhide is at version 40.0.2.  In RHEL
9 it is at version 36.0.1.  Both versions will need to be allowed in
this dependency range.  The backwards-incompatible changes mentioned
upstream between those versions seem to be mostly dropping support for
building against older Python, OpenSSL, and LibreSSL versions.  Other
backwards-incompatible changes relate to already deprecated methods that
do not appear to be in use in this project.